refactor(common): extract duplicated strip proto keys into shared util
