+++ packages/common/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './forward-ref.util.js';
+export * from './strip-proto-keys.util.js';
